+++ Example/Sources/View Controllers/LaunchViewController.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
  
- Copyright (c) 2017-2019 MessageKit
+ Copyright (c) 2017-2022 MessageKit
  
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -28,42 +28,79 @@ import SafariServices
 import SwiftUI
 
 final internal class LaunchViewController: UITableViewController {
-    
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .lightContent
+    private enum Row {
+        case basic, advanced, autocomplete, embedded, customLayout, subview, customInputBar, swiftUI
+        case settings, sourceCode, contributors
+
+        var title: String {
+            switch self {
+            case .basic:
+                return "Basic Example"
+            case .advanced:
+                return "Advanced Example"
+            case .autocomplete:
+                return "Autocomplete Example"
+            case .embedded:
+                return "Embedded Example"
+            case .customLayout:
+                return "Custom Layout Example"
+            case .subview:
+                return "Subview Example"
+            case .customInputBar:
+                return "Custom InputBar Example"
+            case .swiftUI:
+                return "SwiftUI Example"
+            case .settings:
+                return "Settings"
+            case .sourceCode:
+                return "Source Code"
+            case .contributors:
+                return "Contributors"
+            }
+        }
     }
 
-    let cells = ["Basic Example", "Advanced Example", "Autocomplete Example", "Embedded Example", "Custom Layout Example", "Subview Example", "SwiftUI Example", "Settings", "Source Code", "Contributors"]
-    
+    private struct Section {
+        let title: String
+        let rows: [Row]
+    }
+
+    private let sections: [Section] = [
+        .init(title: "Examples", rows: [.basic, .advanced, .autocomplete, .embedded, .customLayout, .subview, .customInputBar, .swiftUI]),
+        .init(title: "Support", rows: [.settings, .sourceCode, .contributors])
+    ]
+
     // MARK: - View Life Cycle
-    
+
+    init() {
+        super.init(style: .insetGrouped)
+    }
+
+    required init?(coder: NSCoder) { nil }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         title = "MessageKit"
         navigationItem.backBarButtonItem = UIBarButtonItem(title: "", style: .plain, target: nil, action: nil)
+        navigationController?.navigationBar.prefersLargeTitles = true
+        navigationController?.navigationBar.tintColor = .primaryColor
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
         tableView.tableFooterView = UIView()
     }
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        navigationController?.navigationBar.prefersLargeTitles = true
-    }
-    
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        navigationController?.navigationBar.prefersLargeTitles = false
-    }
-    
     // MARK: - UITableViewDataSource
 
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        return sections.count
+    }
+
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return cells.count
+        return sections[section].rows.count
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell") ?? UITableViewCell()
-        cell.textLabel?.text = cells[indexPath.row]
+        cell.textLabel?.text = sections[indexPath.section].rows[indexPath.row].title
         cell.accessoryType = .disclosureIndicator
         return cell
     }
@@ -72,51 +109,47 @@ final internal class LaunchViewController: UITableViewController {
     
     // swiftlint:disable cyclomatic_complexity
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let cell = cells[indexPath.row]
+        let cell = sections[indexPath.section].rows[indexPath.row]
         switch cell {
-        case "Basic Example":
+        case .basic:
             let viewController = BasicExampleViewController()
-            let detailViewController = NavigationController(rootViewController: viewController)
+            let detailViewController = UINavigationController(rootViewController: viewController)
             splitViewController?.showDetailViewController(detailViewController, sender: self)
-        case "Advanced Example":
+        case .advanced:
             let viewController = AdvancedExampleViewController()
-            let detailViewController = NavigationController(rootViewController: viewController)
+            let detailViewController = UINavigationController(rootViewController: viewController)
             splitViewController?.showDetailViewController(detailViewController, sender: self)
-        case "Autocomplete Example":
+        case .autocomplete:
             let viewController = AutocompleteExampleViewController()
-            let detailViewController = NavigationController(rootViewController: viewController)
+            let detailViewController = UINavigationController(rootViewController: viewController)
             splitViewController?.showDetailViewController(detailViewController, sender: self)
-        case "Embedded Example":
-            navigationController?.pushViewController(MessageContainerController(), animated: true)
-        case "Custom Layout Example":
-            navigationController?.pushViewController(CustomLayoutExampleViewController(), animated: true)
-        case "SwiftUI Example":
-            if #available(iOS 13, *) {
-                navigationController?.pushViewController(UIHostingController(rootView: SwiftUIExampleView()), animated: true)
-            }
-        case "Settings":
-            let viewController = SettingsViewController()
-            let detailViewController = NavigationController(rootViewController: viewController)
+        case .embedded:
+            splitViewController?.showDetailViewController(MessageContainerController(), sender: self)
+        case .customLayout:
+            splitViewController?.showDetailViewController(CustomLayoutExampleViewController(), sender: self)
+        case .customInputBar:
+            let detailViewController = UINavigationController(rootViewController: CustomInputBarExampleViewController())
             splitViewController?.showDetailViewController(detailViewController, sender: self)
-        case "Subview Example":
+        case .swiftUI:
+            splitViewController?.showDetailViewController(UIHostingController(rootView: SwiftUIExampleView()), sender: self)
+        case .subview:
             let viewController = MessageSubviewContainerViewController()
-            let detailViewController = NavigationController(rootViewController: viewController)
+            let detailViewController = UINavigationController(rootViewController: viewController)
             splitViewController?.showDetailViewController(detailViewController, sender: self)
-        case "Source Code":
-            guard let url = URL(string: "https://github.com/MessageKit/MessageKit") else { return }
-            openURL(url)
-        case "Contributors":
-            guard let url = URL(string: "https://github.com/orgs/MessageKit/teams/contributors/members") else { return }
-            openURL(url)
-        default:
-            assertionFailure("You need to implement the action for this cell: \(cell)")
-            return
+        case .settings:
+            let viewController = SettingsViewController()
+            let detailViewController = UINavigationController(rootViewController: viewController)
+            splitViewController?.showDetailViewController(detailViewController, sender: self)
+        case .sourceCode:
+            openURL(URL(string: "https://github.com/MessageKit/MessageKit")!)
+        case .contributors:
+            openURL(URL(string: "https://github.com/MessageKit/MessageKit/graphs/contributors")!)
         }
     }
     
     func openURL(_ url: URL) {
         let webViewController = SFSafariViewController(url: url)
         webViewController.preferredControlTintColor = .primaryColor
-        splitViewController?.showDetailViewController(webViewController, sender: self)
+        present(webViewController, animated: true)
     }
 }