+++ Sources/Layout/MediaMessageSizeCalculator.swift
@@ -27,8 +27,8 @@ import UIKit
 
 open class MediaMessageSizeCalculator: MessageSizeCalculator {
 
-    open override func messageContainerSize(for message: MessageType) -> CGSize {
-        let maxWidth = messageContainerMaxWidth(for: message)
+    open override func messageContainerSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
+        let maxWidth = messageContainerMaxWidth(for: message, at: indexPath)
         let sizeForMediaItem = { (maxWidth: CGFloat, item: MediaItem) -> CGSize in
             if maxWidth < item.size.width {
                 // Maintain the ratio if width is too great